# services/order-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: Order management microservice for e-commerce platform
  version: 1.0.0
  contact:
    name: Eren
    email: eren@example.com

servers:
  - url: http://localhost:8003
    description: Local development
  - url: http://order-service:8003
    description: Kubernetes cluster

tags:
  - name: Health
    description: Health check endpoints
  - name: Orders
    description: Order management operations
  - name: User Orders
    description: User-specific order operations

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if service is healthy
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ready:
    get:
      tags:
        - Health
      summary: Readiness check
      description: Check if service is ready to accept requests
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is not ready

  /api/orders:
    post:
      tags:
        - Orders
      summary: Create new order
      description: Create a new order with order items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid request body
        '500':
          description: Internal server error

    get:
      tags:
        - Orders
      summary: List all orders
      description: Get paginated list of all orders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

  /api/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve a specific order by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found

  /api/orders/{id}/status:
    patch:
      tags:
        - Orders
      summary: Update order status
      description: Update the status of an existing order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Status updated successfully
        '400':
          description: Invalid status
        '404':
          description: Order not found

  /api/orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel order
      description: Cancel an existing order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled successfully
        '404':
          description: Order not found

  /api/users/{user_id}/orders:
    get:
      tags:
        - User Orders
      summary: Get user's orders
      description: Get paginated list of orders for a specific user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object

    CreateOrderRequest:
      type: object
      required:
        - user_id
        - shipping_address
        - order_items
      properties:
        user_id:
          type: string
        shipping_address:
          type: string
        billing_address:
          type: string
        notes:
          type: string
        order_items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItemRequest'

    CreateOrderItemRequest:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
        quantity:
          type: integer
          minimum: 1

    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]

    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        total_amount:
          type: number
          format: float
        currency:
          type: string
        shipping_address:
          type: string
        billing_address:
          type: string
        notes:
          type: string
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        product_id:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
          format: float
        subtotal:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OrderResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          $ref: '#/components/schemas/Order'

    OrderListResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'

    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer